SmartMilesPy/requirements.txt
appdirs==1.4.3
Babel==2.4.0
Django==1.11.1
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
packaging==16.8
phonenumberslite==8.4.3
pip==9.0.1
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
setuptools==35.0.2
six==1.10.0
